% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunction.R
\name{myfunction}
\alias{myfunction}
\title{My customized function
Output the column names, types, and the five most common elements.}
\usage{
myfunction(x)
}
\arguments{
\item{x}{A dataframe or dataset}
}
\description{
My customized function
Output the column names, types, and the five most common elements.
}
\examples{
dataframe <- matrix(1:99)
myfunction(dataframe)
}
\author{
Hejing Wang Wang Wang
}
