% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5.R
\name{pythagorean}
\alias{pythagorean}
\title{Calculates the third side of a right triangle based on the Pythagorean Theorem.}
\usage{
pythagorean(a = NULL, b = NULL, c = NULL)
}
\arguments{
\item{a, b, c}{numeric}
}
\description{
Calculates the length of the third side of a right triangle when two sides are inputted. c is the hypotenuse, a is the height, and b is the base.
}
\examples{
pythagorean(a = 3 , b=4)
pythagorean(a=3 , c=5)
}
\author{
Rami Hawila
}
