% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5.R
\name{trimmed_mean}
\alias{trimmed_mean}
\title{Calculates the mean of the vector when s smallest elements and l largest elements are removed from the vector.}
\usage{
trimmed_mean(x, s, l)
}
\arguments{
\item{x}{numeric vector}

\item{s}{number of smallest elements removed.}

\item{l}{number of largest elements removed.}
}
\description{
Calculates the mean of the vector when s smallest elements and l largest elements are removed from the vector.
}
\examples{
trimmed_mean(c(1,2,3,4,5,6),1,2)
trimmed_mean(c(1,7,3,2,5,0.5,9,10),1,2)
}
\author{
Rami Hawila
}
